+++ tests/Magick.NET.Tests/Pixels/SafePixelCollectionTests/TheSetByteAreaMethod.cs
@@ -47,7 +47,7 @@ public partial class SafePixelCollectionTests
             var values = new byte[113 * 108 * image.ChannelCount];
             pixels.SetByteArea(10, 10, 113, 108, values);
 
-            ColorAssert.Equal(MagickColors.Black, image, image.Width - 1, image.Height - 1);
+            ColorAssert.Equal(MagickColors.Black, image, (int)image.Width - 1, (int)image.Height - 1);
         }
 
         [Fact]
@@ -67,7 +67,7 @@ public partial class SafePixelCollectionTests
             var values = new byte[113 * 108 * image.ChannelCount];
             pixels.SetByteArea(new MagickGeometry(10, 10, 113, 108), values);
 
-            ColorAssert.Equal(MagickColors.Black, image, image.Width - 1, image.Height - 1);
+            ColorAssert.Equal(MagickColors.Black, image, (int)(int)image.Width - 1, (int)image.Height - 1);
         }
     }
 }
